fix(gateway): preserve explicit session labels in derived titles
