+++ src/gateway/session-utils.test.ts
@@ -76,13 +76,6 @@ function createModelDefaultsConfig(params: {
   } as OpenClawConfig;
 }
 
-function requireString(value: string | undefined, label: string): string {
-  if (!value) {
-    throw new Error(`expected ${label}`);
-  }
-  return value;
-}
-
 describe("gateway session utils", () => {
   afterEach(() => {
     resetConfigRuntimeState();
@@ -1753,96 +1746,9 @@ describe("resolveSessionDisplayModelIdentityRef", () => {
   });
 });
 
-describe("deriveSessionTitle", () => {
-  test("returns undefined for undefined entry", () => {
-    expect(deriveSessionTitle(undefined)).toBeUndefined();
-  });
-
-  test("prefers displayName when set", () => {
-    const entry = {
-      sessionId: "abc123",
-      updatedAt: Date.now(),
-      displayName: "My Custom Session",
-      subject: "Group Chat",
-    } as SessionEntry;
-    expect(deriveSessionTitle(entry)).toBe("My Custom Session");
-  });
-
-  test("falls back to subject when displayName is missing", () => {
-    const entry = {
-      sessionId: "abc123",
-      updatedAt: Date.now(),
-      subject: "Dev Team Chat",
-    } as SessionEntry;
-    expect(deriveSessionTitle(entry)).toBe("Dev Team Chat");
-  });
-
-  test("uses first user message when displayName and subject missing", () => {
-    const entry = {
-      sessionId: "abc123",
-      updatedAt: Date.now(),
-    } as SessionEntry;
-    expect(deriveSessionTitle(entry, "Hello, how are you?")).toBe("Hello, how are you?");
-  });
-
-  test("truncates long first user message to 60 chars with ellipsis", () => {
-    const entry = {
-      sessionId: "abc123",
-      updatedAt: Date.now(),
-    } as SessionEntry;
-    const longMsg =
-      "This is a very long message that exceeds sixty characters and should be truncated appropriately";
-    const result = requireString(deriveSessionTitle(entry, longMsg), "truncated session title");
-    expect(result.length).toBeLessThanOrEqual(60);
-    expect(result.endsWith("…")).toBe(true);
-  });
-
-  test("truncates at word boundary when possible", () => {
-    const entry = {
-      sessionId: "abc123",
-      updatedAt: Date.now(),
-    } as SessionEntry;
-    const longMsg = "This message has many words and should be truncated at a word boundary nicely";
-    const result = requireString(deriveSessionTitle(entry, longMsg), "word-boundary session title");
-    expect(result.endsWith("…")).toBe(true);
-    expect(result.includes("  ")).toBe(false);
-  });
-
-  test("falls back to sessionId prefix with date", () => {
-    const entry = {
-      sessionId: "abcd1234-5678-90ef-ghij-klmnopqrstuv",
-      updatedAt: new Date("2024-03-15T10:30:00Z").getTime(),
-    } as SessionEntry;
-    const result = deriveSessionTitle(entry);
-    expect(result).toBe("abcd1234 (2024-03-15)");
-  });
-
-  test("falls back to sessionId prefix without date when updatedAt missing", () => {
-    const entry = {
-      sessionId: "abcd1234-5678-90ef-ghij-klmnopqrstuv",
-      updatedAt: 0,
-    } as SessionEntry;
-    const result = deriveSessionTitle(entry);
-    expect(result).toBe("abcd1234");
-  });
-
-  test("trims whitespace from displayName", () => {
-    const entry = {
-      sessionId: "abc123",
-      updatedAt: Date.now(),
-      displayName: "  Padded Name  ",
-    } as SessionEntry;
-    expect(deriveSessionTitle(entry)).toBe("Padded Name");
-  });
-
-  test("ignores empty displayName and falls through", () => {
-    const entry = {
-      sessionId: "abc123",
-      updatedAt: Date.now(),
-      displayName: "   ",
-      subject: "Actual Subject",
-    } as SessionEntry;
-    expect(deriveSessionTitle(entry)).toBe("Actual Subject");
+describe("deriveSessionTitle compatibility re-export", () => {
+  test("deriveSessionTitle is re-exported from session-utils", () => {
+    expect(typeof deriveSessionTitle).toBe("function");
   });
 });
