+++ src/gateway/session-utils.ts
@@ -91,6 +91,7 @@ import {
   readSessionTitleFieldsFromTranscriptAsync,
   readSessionTitleFieldsFromTranscript,
 } from "./session-utils.fs.js";
+import { deriveSessionTitle } from "./session-title.js";
 import type {
   GatewayAgentRow,
   GatewaySessionRow,
@@ -123,6 +124,7 @@ export {
 } from "./session-utils.fs.js";
 export type { ReadSessionMessagesAsyncOptions } from "./session-utils.fs.js";
 export { canonicalizeSpawnedByForAgent, resolveSessionStoreKey } from "./session-store-key.js";
+export { deriveSessionTitle } from "./session-title.js";
 export type {
   GatewayAgentRow,
   GatewaySessionRow,
@@ -132,8 +134,6 @@ export type {
   SessionsPreviewEntry,
   SessionsPreviewResult,
 } from "./session-utils.types.js";
-
-const DERIVED_TITLE_MAX_LEN = 60;
 
 function tryResolveExistingPath(value: string): string | null {
   try {
@@ -189,56 +189,6 @@ function resolveIdentityAvatarUrl(
   } catch {
     return undefined;
   }
-}
-
-function formatSessionIdPrefix(sessionId: string, updatedAt?: number | null): string {
-  const prefix = sessionId.slice(0, 8);
-  if (updatedAt && updatedAt > 0) {
-    const d = new Date(updatedAt);
-    const date = d.toISOString().slice(0, 10);
-    return `${prefix} (${date})`;
-  }
-  return prefix;
-}
-
-function truncateTitle(text: string, maxLen: number): string {
-  if (text.length <= maxLen) {
-    return text;
-  }
-  const cut = text.slice(0, maxLen - 1);
-  const lastSpace = cut.lastIndexOf(" ");
-  if (lastSpace > maxLen * 0.6) {
-    return cut.slice(0, lastSpace) + "…";
-  }
-  return cut + "…";
-}
-
-export function deriveSessionTitle(
-  entry: SessionEntry | undefined,
-  firstUserMessage?: string | null,
-): string | undefined {
-  if (!entry) {
-    return undefined;
-  }
-
-  if (normalizeOptionalString(entry.displayName)) {
-    return normalizeOptionalString(entry.displayName);
-  }
-
-  if (normalizeOptionalString(entry.subject)) {
-    return normalizeOptionalString(entry.subject);
-  }
-
-  if (firstUserMessage?.trim()) {
-    const normalized = firstUserMessage.replace(/\s+/g, " ").trim();
-    return truncateTitle(normalized, DERIVED_TITLE_MAX_LEN);
-  }
-
-  if (entry.sessionId) {
-    return formatSessionIdPrefix(entry.sessionId, entry.updatedAt);
-  }
-
-  return undefined;
 }
 
 function resolveSessionRuntimeMs(
